Add book's cover upload function to the book service and book creation form. Image files are saved using Firebase storage and displayed in the corresponding "single book view page".

+++ src/app/book-list/book-list.component.ts
@@ -32,7 +32,9 @@ export class BookListComponent implements OnInit, OnDestroy{
   }
 
   onDeleteBook(book: Book){
-  	this.booksService.removeBook(book);
+    if(confirm('"'+book.title+'"\n\rCe livre sera supprimé de votre bibliothèque. Continuer ?')){
+      this.booksService.removeBook(book);
+    }
   }
 
   onNewBook(){
